ISU for Baytrail(1.2.7)
BZ:134315

isu_byt is ISU for Baytrail. It has linux and windows versions for ISU
with accompanying libraries.
The config folder contains the sample configuration files that are required for
IAFW/PDR signing. Secure_Boot_Tools.pptx contains BKM for ISU
Updated the version to 1.2.7. This version of ISU should work on both 10.04 and 12.04
Also added OS Manifest fix (Payload size in the manifest)

Change-Id: Iedc5d9fbf7da50ac51a6316ca26627f9045747f1
Signed-off-by: Khushboo Bindlish <khushboo.bindlish@intel.com>