ISU 1.2.7 for Baytrail
BZ:133818

1.ISU will now sign the hash and size of the OS image instead of the whole image.
This is to be consistent with Cloverview infrastructure.
During an AOS build, signing is executed on a signing server that requires the file
to be copied to the server.
So, instead of copying the whole image, it will be more convenient to copy the images hash
and size and sign using that.
So,
./isu   -p 2   -t 6  Takes the hash and size
./isu   -p 2   -t 4  Takes the whole image (i.e. no_osip_boot.bin)
The hash file contains the sha256sum of the os binary image and the size in hex.
Whenever the OS image changes, the hash will have to be updated in this file.
2. Provide the public key for OS Signing as an input.
For previous versions, ISU would only take private key as an input and derive the
public key from that. I have changed it to also take public key as an input.
So that the manifest can be signed on a separate signing server and the end user
does not need the private key if HSM or some remote signing is used.
Please note that if you provide only the private key, it will still attempt to derive
the public key. So, your scripts dont have to change for this.

Change-Id: If5363b518595115ee9b2dea46ada249346d36f24
Signed-off-by: Khushboo Bindlish <khushboo.bindlish@intel.com>